+++ utils/dom.ts
@@ -1,0 +1,11 @@
+export const offsetLeft = (element: any) => {
+  let el = element
+  let x = el.offsetLeft
+
+  while (el.offsetParent) {
+    x += el.offsetParent.offsetLeft
+    el = el.offsetParent
+  }
+
+  return x
+}
